+++ src/codeql.ts
@@ -105,6 +105,10 @@ export interface CodeQL {
     threadsFlag: string,
     automationDetailsId: string | undefined
   ): Promise<string>;
+  /**
+   * Run 'codeql database cleanup'.
+   */
+  databaseCleanup(databasePath: string): Promise<void>;
 }
 
 export interface ResolveLanguagesOutput {
@@ -485,6 +489,7 @@ export function setCodeQL(partialCodeql: Partial<CodeQL>): CodeQL {
     resolveLanguages: resolveFunction(partialCodeql, "resolveLanguages"),
     resolveQueries: resolveFunction(partialCodeql, "resolveQueries"),
     databaseAnalyze: resolveFunction(partialCodeql, "databaseAnalyze"),
+    databaseCleanup: resolveFunction(partialCodeql, "databaseCleanup"),
   };
   return cachedCodeQL;
 }
@@ -750,6 +755,10 @@ function getCodeQLForCmd(cmd: string): CodeQL {
         },
       }).exec();
       return output;
+    },
+    async databaseCleanup(databasePath: string): Promise<void> {
+      const args = ["database", "cleanup", databasePath];
+      await new toolrunner.ToolRunner(cmd, args).exec();
     },
   };
 }